Optionally loosen strictness for certain fields (#11)
Meet Manager fairly regards certain fields as optional which are listed
as M1 in the published spec. We should support serializing and
deserializing SDIF files where these values are absent.

Adds a new strict mode that requires we match the spec exactly, which is
off by default. Notes on the RelayName model that these fields are
Optional even though they're "m1".

Fixes #9.